chore: Refactored reducer logic to separate file to clean things up
