+++ app/packages/frontend/pages/index.tsx
@@ -2,9 +2,9 @@ import { Box, Button, Heading, Input, Text } from '@chakra-ui/react'
 import { ChainId, useEthers, useSendTransaction } from '@usedapp/core'
 import { ethers, providers, utils } from 'ethers'
 import React, { useReducer } from 'react'
-import Poster from '@poster/contracts/artifacts/contracts/Poster.sol/Poster.json'
+
 import Layout from '../components/layout/Layout'
-import { Poster as PosterType } from '@poster/contracts/typechain'
+import { initialState, reducer, setPostContent } from '../lib/reducers'
 
 /**
  * Constants & Helpers
@@ -15,49 +15,7 @@ const localProvider = new providers.StaticJsonRpcProvider(
 )
 
 // Single address across all networks due to Singleton Factory pattern
-const CONTRACT_ADDRESS = '0xA0c7A49916Ce3ed7dd15871550212fcc7079AD61'
-
-/**
- * Prop Types
- */
-type StateType = {
-  inputValue: string
-  isLoading: boolean
-}
-type ActionType =
-  | {
-      type: 'SET_INPUT_VALUE'
-      inputValue: StateType['inputValue']
-    }
-  | {
-      type: 'SET_LOADING'
-      isLoading: StateType['isLoading']
-    }
-
-/**
- * Component
- */
-const initialState: StateType = {
-  inputValue: '',
-  isLoading: false,
-}
-
-function reducer(state: StateType, action: ActionType): StateType {
-  switch (action.type) {
-    case 'SET_INPUT_VALUE':
-      return {
-        ...state,
-        inputValue: action.inputValue,
-      }
-    case 'SET_LOADING':
-      return {
-        ...state,
-        isLoading: action.isLoading,
-      }
-    default:
-      throw new Error()
-  }
-}
+const POSTER_CONTRACT_ADDRESS = '0xA0c7A49916Ce3ed7dd15871550212fcc7079AD61'
 
 function HomeIndex(): JSX.Element {
   const [state, dispatch] = useReducer(reducer, initialState)
@@ -70,33 +28,6 @@ function HomeIndex(): JSX.Element {
   const { sendTransaction } = useSendTransaction({
     signer: localProvider.getSigner(),
   })
-
-  // call the smart contract, send an update
-  async function setPostContent() {
-    if (!state.inputValue) return
-    if (library) {
-      dispatch({
-        type: 'SET_LOADING',
-        isLoading: true,
-      })
-      const signer = library.getSigner()
-      const contract = new ethers.Contract(
-        CONTRACT_ADDRESS,
-        Poster.abi,
-        signer
-      ) as unknown as PosterType
-      const transaction = await contract.post(state.inputValue)
-      await transaction.wait()
-      dispatch({
-        type: 'SET_LOADING',
-        isLoading: false,
-      })
-      dispatch({
-        type: 'SET_INPUT_VALUE',
-        inputValue: ''
-      })
-    }
-  }
 
   function sendFunds(): void {
     sendTransaction({
@@ -132,7 +63,9 @@ function HomeIndex(): JSX.Element {
             mt="2"
             colorScheme="teal"
             isLoading={state.isLoading}
-            onClick={setPostContent}
+            onClick={() =>
+              setPostContent(POSTER_CONTRACT_ADDRESS, state, library, dispatch)
+            }
           >
             Post
           </Button>
